First Commit to this branch

+++ UserDatabaseJSON.cs
@@ -12,6 +12,11 @@ namespace AdsSolution
 {
     public class UserDatabaseJSON
     {
+        public static int LoginToken
+        {
+            get;
+            set;
+        }
         private string filePath
         {
             get;
@@ -23,45 +28,38 @@ namespace AdsSolution
             if (!DatabaseExists()) DatabaseCreate();
 
         }
+       
         private bool DatabaseExists()
         {
-
             return File.Exists(Path.Combine(filePath, "Users.json"));
         }
         private void DatabaseCreate()
         {
-            if (!Directory.Exists(filePath))
-                Directory.CreateDirectory(filePath);
+            if (!Directory.Exists(filePath)) Directory.CreateDirectory(filePath);
 
-            if (!File.Exists(Path.Combine(filePath, "Users.json")))
-                File.Create(Path.Combine(filePath, "Users.json"));
+            if (!File.Exists(Path.Combine(filePath, "Users.json"))) File.Create(Path.Combine(filePath, "Users.json"));
         }
-        public User LoadUser(int LoginToken)
-        {
-            var Text = File.ReadAllText(Path.Combine(filePath, "Users.json"));
-            var UserList = JsonConvert.DeserializeObject<Dictionary<int,User>>(Text);
-            return UserList[LoginToken];
-        }
-    
         public void CreateNewUser(User U)
         {
-            var UserList = new Dictionary<int, User>();
+            var UserList = new Dictionary<int,
+            User>();
             var Text = File.ReadAllText(Path.Combine(filePath, "Users.json"));
             if (!string.IsNullOrEmpty(Text))
             {
-                UserList = JsonConvert.DeserializeObject<Dictionary<int, User>>(Text);
+                UserList = JsonConvert.DeserializeObject<Dictionary<int,User>>(Text);
             }
             UserList.Add(U.GetHashCode(), U);
             File.WriteAllText(Path.Combine(filePath, "Users.json"), JsonConvert.SerializeObject(UserList));
         }
-        public int GetLoginToken(string email, string password)
+        public void LoginUser(string email, string password)
         {
-
             var Text = File.ReadAllText(Path.Combine(filePath, "Users.json"));
-            var UserList = JsonConvert.DeserializeObject<Dictionary<int, User>>(Text);
-            return UserList.FirstOrDefault(x => (String.Equals(x.Value.Email, email) && String.Equals(x.Value.Password, password))).Key;
-            
+            var UserList = JsonConvert.DeserializeObject<Dictionary<int,
+            User>>(Text);
+            var m = UserList.FirstOrDefault(x => (String.Equals(x.Value.Email, email) && String.Equals(x.Value.Password, password))).Key;
+            //   LoadUser();
+            ParentForm P = new ParentForm(UserList[m]);
+            P.Show();
         }
-        
     }
 }
